+++ CHANGELOG.md
@@ -12,7 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added the ability to modify port settings whilst the port is open (port is quickly disconnected and reconnected with new settings).
 - Added support for the ESC[1J, ESC[2J and ESC[3J "Erase in Display" ASCII escape codes (ESC[0J was already supported).
 - Added a "Manual" page to the app, which contains the start of a user manual for NinjaTerm.
-- Added unit tests to make sure app data in local storage can be upgraded correctly. 
+- Added unit tests to make sure app data in local storage can be upgraded correctly.
 
 ### Fixed
 
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Changed
 
 - Switched from a CJS to ESM build in vite. CJS build is deprecated. See https://vitejs.dev/guide/troubleshooting.html#vite-cjs-node-api-deprecated for more info.
+- Cursor up escape codes cannot move the cursor up into the scrollback buffer to correctly emulate terminal behaviour.
 
 ## [4.18.0] - 2024-06-16